refactor(tools): share the DataHandler mechanism between the two writers
SonarCloud's quality gate failed the pull request on
new_duplicated_lines_density 5.3% against a 3% threshold, locating 303
duplicated lines between the two writing tools.

That is the consequence of this branch's own argument rather than a
reversal of it: if the duplication really is mechanism, and everything
that DECIDES something differs, then the mechanism is extractable without
building an abstraction over decisions. Sharing boilerplate is not the
same as unifying behaviour.

A trait, not a base class -- following CollectsEnvironmentTrait and
ResolvesLanguageLabelTrait. A base class would open exactly the
inheritance axis this branch argued against.

Shared, because it only executes: the TCA/LANG/BE_USER environment check,
the live-workspace guard, the bounded DataHandler error summary, the TCA
column lookup, two formatting helpers. ADR-136 had already argued both
guards are process-level rather than record-level, which is what makes
them safe to share.

Left per tool, with reasons in the ADR: the neutral refusal strings (each
is paired with the READ tool of the same records; one shared string would
break that pairing), the four declaration methods (a trait answering them
would make a third writer INHERIT being non-admin and idempotent instead
of deciding it -- the largest chunk deliberately not taken), the read-back
check, the row lookup, the field allow-list, the language rule, the
authorisation.

The sha1() fixture lines flagged as php:S4790 are untouched: they are
TYPO3 lookup keys, not a security context -- AbstractDriver::hashIdentifier()
is literally sha1(), and FileIndexRepository queries the column by it.
Marked false positive in SonarCloud with that evidence.
